# ARROW-7948: [Go] Decimal128 Integration fix