--- a/dev/archery/archery/integration/datagen.py
+++ b/dev/archery/archery/integration/datagen.py
@@ -1527,7 +1527,7 @@ def get_generated_json_files(tempdir=None):
         .skip_category('Go'),  # TODO(ARROW-7901)
 
         generate_decimal128_case()
-        .skip_category('Go')  # TODO(ARROW-7948): Decimal + Go
+        # .skip_category('Go')  # TODO(ARROW-7948): Decimal + Go
         .skip_category('Rust'),
 
         generate_decimal256_case()