--- a/go/arrow/go.mod
+++ b/go/arrow/go.mod
@@ -20,12 +20,11 @@ go 1.12
 
 require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
-	github.com/frankban/quicktest v1.11.3 // indirect
 	github.com/golang/protobuf v1.4.2
 	github.com/google/flatbuffers v1.11.0
+	github.com/google/go-cmp v0.5.4 // indirect
 	github.com/klauspost/compress v1.11.13
-	github.com/pierrec/lz4 v2.6.0+incompatible
-	github.com/pierrec/lz4/v4 v4.1.4 // indirect
+	github.com/pierrec/lz4/v4 v4.1.4
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/stretchr/testify v1.2.0
 	golang.org/x/net v0.0.0-20200904194848-62affa334b73 // indirect