Had to redesign a bit of the arrjson handling to do this properly, but I think it's better this way anyways.