@zeroshade 2021-04-20
tagging @pitrou since the issue ARROW-7948 was originally filed by him.

CC @sbinet @emkornfield