@emkornfield 2021-04-22
Two questions, otherwise looks reasonable to me.