@zeroshade 2021-04-23
the slight changes in the updated commits significantly improve the performance which can be seen by the benchmarks.

Before:
```
goos: windows
goarch: amd64
pkg: github.com/apache/arrow/go/arrow/decimal128
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkBigIntToDecimal-12    	20276983	        58.21 ns/op	      48 B/op	       1 allocs/op
BenchmarkDecimalToBigInt-12    	 8945876	       141.2 ns/op	     112 B/op	       4 allocs/op
PASS
ok  	github.com/apache/arrow/go/arrow/decimal128	2.839s
```
After:
```
goos: windows
goarch: amd64
pkg: github.com/apache/arrow/go/arrow/decimal128
cpu: Intel(R) Core(TM) i7-8850H CPU @ 2.60GHz
BenchmarkBigIntToDecimal-12    	262177144	         4.549 ns/op	       0 B/op	       0 allocs/op
BenchmarkDecimalToBigInt-12    	21051966	        58.65 ns/op	      48 B/op	       2 allocs/op
PASS
ok  	github.com/apache/arrow/go/arrow/decimal128	3.045s
```