@emkornfield 2021-04-27
+1.  Thank you.  @zeroshade not sure if you are interested but I don't think Go supports Decimal256 yet.