@zeroshade 2021-04-27
@emkornfield It doesn't, that's on my list of stuff that i plan on making PRs for haha. 😄